requirements.txt
# 数据库
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# 网络和解析
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
requests-cache==1.1.0
aiohttp==3.9.1
urllib3==2.0.7

# 爬虫相关
robotexclusionrulesparser==1.7.1
cssselect==1.2.0

# 工具
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3
pydantic==2.5.2
pydantic-settings==2.1.0

# 测试
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
responses==0.24.1 